refactor[landing]: use github api to fetch stars instead of hardcoding
